Add proceso de aprobacion de Adm Solicitu

+++ Convertidor/Data/Entities/Presupuesto/PRE_DETALLE_COMPROMISOS.cs
@@ -5,16 +5,16 @@
         public int CODIGO_DETALLE_COMPROMISO { get; set; }
         public int CODIGO_COMPROMISO { get; set; }
         public int CODIGO_DETALLE_SOLICITUD { get; set; }
-        public int CANTIDAD { get; set; }
-        public int CANTIDAD_ANULADA { get; set; }
+        public decimal CANTIDAD { get; set; }
+        public decimal CANTIDAD_ANULADA { get; set; }
         public int UDM_ID { get; set; }
         public string DESCRIPCION { get; set; } = string.Empty;
-        public int PRECIO_UNITARIO { get; set; }
-        public int POR_DESCUENTO { get; set; }
-        public int MONTO_DESCUENTO { get; set; }
+        public decimal PRECIO_UNITARIO { get; set; }
+        public decimal POR_DESCUENTO { get; set; }
+        public decimal MONTO_DESCUENTO { get; set; }
         public int TIPO_IMPUESTO_ID { get; set; }
-        public int POR_IMPUESTO { get; set; }
-        public int MONTO_IMPUESTO { get; set; }
+        public decimal POR_IMPUESTO { get; set; }
+        public decimal MONTO_IMPUESTO { get; set; }
         public string EXTRA1 { get; set; } = string.Empty;
         public string EXTRA2 { get; set; } = string.Empty;
         public string EXTRA3 { get; set; } = string.Empty;
